Make library usable outside of the repository.
So others can use it :D.
