+++ .gitignore
@@ -1,7 +1,6 @@
 # Binary file
-sshaws
-sshaws.mac
-
+/sshaws
+/sshaws.mac
 
 # SSH Stuff
 travisci/github_deploy_key
