+++ go.mod
@@ -2,4 +2,11 @@ module sshaws
 
 go 1.12
 
-require github.com/aws/aws-sdk-go v1.34.18
+require (
+	github.com/aws/aws-sdk-go v1.34.18
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/kr/pretty v0.1.0 // indirect
+	golang.org/x/text v0.3.2 // indirect
+	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
+	gopkg.in/yaml.v2 v2.2.4 // indirect
+)
