+++ pkg/cmd/login/helpers.go
@@ -1,0 +1,19 @@
+package login
+
+//Instance struct
+type Instance struct {
+	Name string
+	IP   string
+	ID   string
+	Size string
+}
+
+//NewInstance creator
+func NewInstance(name, ip, id, size string) *Instance {
+	newInstance := new(Instance)
+	newInstance.Name = name
+	newInstance.IP = ip
+	newInstance.ID = id
+	newInstance.Size = size
+	return newInstance
+}
