+++ pkg/cmd/login/login.go
@@ -1,20 +1,15 @@
-package main
+package login
 
-import (
-	"sshaws/helpers"
-	"sshaws/params"
-)
-
-func main() {
-	config := params.Read()
-	env, app, name, region, user, silent, ssh := helpers.ReturnConfiguration(config)
+func NewLogin(env string, app string, name string, region string, user string, silent bool, ssh bool) {
 	rawInstanceList := filterInstances(region, env, app, name, silent, ssh)
 	instances := getInstancesInfo(rawInstanceList)
+
 	if silent {
 		showIPsList(instances)
 	} else {
 		showInstanceList(instances, user)
 	}
+
 	selectedInstance := selectInstanceIndex(instances)
 
 	if ssh {
