+++ pkg/cmd/login/ssh.go
@@ -1,4 +1,4 @@
-package main
+package login
 
 import (
 	"fmt"
